loosy convertion

+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/PDFView.java
@@ -280,7 +280,7 @@ public class PDFView extends RelativeLayout {
     public PointF mapDeviceCoordsToPage(int pageIndex, int startX, int startY, int sizeX,
                                            int sizeY, int rotate, int deviceX, int deviceY) {
 
-      int startYWithSpacing = startY + pdfFile.getPageSpacing(pageIndex, 1);
+      int startYWithSpacing = startY + (int)pdfFile.getPageSpacing(pageIndex, 1);
       Log.d("offset pixel", String.valueOf(startYWithSpacing));
       return pdfFile.mapDeviceCoordsToPage(pageIndex, startX, startYWithSpacing, sizeX, sizeY, rotate, deviceX, deviceY);
     }
